Perbaikan keamanan user dan penambahan fitur upload cover foto

+++ sosling/app/Http/Controllers/PostsController.php
@@ -3,11 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -49,8 +60,26 @@ class PostsController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'price' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //Handle File yang diupload
+        if($request->hasFile('cover_image')){
+            //Get filename dengan extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         // Buat Post
         $post = new Post;
@@ -58,6 +87,7 @@ class PostsController extends Controller
         $post->price = $request->input('price');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -84,6 +114,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //Cek identitas user
+        if(auth()->user()->id !==$post->user_id){
+            return view('/posts')->with('error', 'Anda tidak bisa mengedit penawaran orang lain');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -102,11 +138,29 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
+        //Handle File yang diupload
+        if($request->hasFile('cover_image')){
+            //Get filename dengan extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }
+
         // Buat Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->price = $request->input('price');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Postingan Berhasil Diupdate');
@@ -122,9 +176,16 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //Cek identitas user
+        if(auth()->user()->id !==$post->user_id){
+            return view('/posts')->with('error', 'Anda tidak bisa menghapus penawaran orang lain');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            //Hapus Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Postingan Berhasil Dihapus');
-
-
     }
 }
